+++ .changeset/khaki-rocks-wave.md
@@ -1,5 +1,6 @@
 ---
 "@ledgerhq/coin-solana": minor
+"@ledgerhq/live-common": minor
 ---
 
 support RAW tx for Solana
