--- libs/coin-modules/coin-solana/src/helpers/kind/mappers/stakeCreateAccount.ts
@@ -1,6 +1,0 @@
-import { OperationTypeMapper } from "../type";
-
-export const mapper: OperationTypeMapper = {
-  kind: "stake.createAccount",
-  operationType: "DELEGATE",
-};
