--- libs/coin-modules/coin-solana/src/helpers/kind/mappers/stakeDelagate.test.ts
@@ -1,8 +1,0 @@
-import { mapper } from "./stakeDelegate";
-
-describe("Testing stakeDelegate mapper", () => {
-  it("should map the kind stake.delegate to the operation type DELEGATE", () => {
-    expect(mapper.kind).toEqual("stake.delegate");
-    expect(mapper.operationType).toEqual("DELEGATE");
-  });
-});
