--- libs/coin-modules/coin-solana/src/helpers/kind/mappers/stakeSplit.test.ts
@@ -1,8 +1,0 @@
-import { mapper } from "./stakeSplit";
-
-describe("Testing stakeSplit mapper", () => {
-  it("should map the kind stake.split to the operation type OUT", () => {
-    expect(mapper.kind).toEqual("stake.split");
-    expect(mapper.operationType).toEqual("OUT");
-  });
-});
