--- libs/coin-modules/coin-solana/src/helpers/kind/mappers/stakeUndelegate.test.ts
@@ -1,8 +1,0 @@
-import { mapper } from "./stakeUndelegate";
-
-describe("Testing stakeUndelegate mapper", () => {
-  it("should map the kind stake.undelegate to the operation type UNDELEGATE", () => {
-    expect(mapper.kind).toEqual("stake.undelegate");
-    expect(mapper.operationType).toEqual("UNDELEGATE");
-  });
-});
