--- libs/coin-modules/coin-solana/src/helpers/kind/mappers/stakeWithdraw.test.ts
@@ -1,8 +1,0 @@
-import { mapper } from "./stakeWithdraw";
-
-describe("Testing stakeWithdraw mapper", () => {
-  it("should map the kind stake.withdraw to the operation type IN", () => {
-    expect(mapper.kind).toEqual("stake.withdraw");
-    expect(mapper.operationType).toEqual("IN");
-  });
-});
