--- libs/coin-modules/coin-solana/src/helpers/kind/mappers/stakeWithdraw.ts
@@ -1,7 +1,0 @@
-import { OperationType } from "@ledgerhq/types-live";
-import { OperationTypeMapper } from "../type";
-
-export const mapper: OperationTypeMapper = {
-  kind: "stake.withdraw",
-  operationType: "IN" as OperationType,
-};
