--- libs/coin-modules/coin-solana/src/helpers/kind/mappers/tokenCreateATA.test.ts
@@ -1,8 +1,0 @@
-import { mapper } from "./tokenCreateATA";
-
-describe("Testing tokenCreateATA mapper", () => {
-  it("should map the kind token.createATA to the operation type OPT_IN", () => {
-    expect(mapper.kind).toEqual("token.createATA");
-    expect(mapper.operationType).toEqual("OPT_IN");
-  });
-});
