--- libs/coin-modules/coin-solana/src/helpers/kind/mappers/tokenTransfer.test.ts
@@ -1,8 +1,0 @@
-import { mapper } from "./tokenTransfer";
-
-describe("Testing tokenTransfer mapper", () => {
-  it("should map the kind token.transfer to the operation type FEES", () => {
-    expect(mapper.kind).toEqual("token.transfer");
-    expect(mapper.operationType).toEqual("FEES");
-  });
-});
