--- libs/coin-modules/coin-solana/src/helpers/kind/mappers/tokenTransfer.ts
@@ -1,7 +1,0 @@
-import { OperationType } from "@ledgerhq/types-live";
-import { OperationTypeMapper } from "../type";
-
-export const mapper: OperationTypeMapper = {
-  kind: "token.transfer",
-  operationType: "FEES" as OperationType,
-};
