--- libs/coin-modules/coin-solana/src/helpers/kind/mappers/transfer.test.ts
@@ -1,8 +1,0 @@
-import { mapper } from "./transfer";
-
-describe("Testing transfer mapper", () => {
-  it("should map the kind transfer to the operation type OUT", () => {
-    expect(mapper.kind).toEqual("transfer");
-    expect(mapper.operationType).toEqual("OUT");
-  });
-});
