--- libs/coin-modules/coin-solana/src/helpers/kind/type.ts
@@ -1,6 +1,0 @@
-import { OperationType } from "@ledgerhq/types-live";
-
-export type OperationTypeMapper = {
-  kind: string;
-  operationType: OperationType;
-};
